backend/pyproject.toml
[project]
name = "boardy-backend"
version = "0.1.0"
description = "Boardy — Claude-powered Kanban SaaS"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "fastmcp>=2.0.0",
    "sqlmodel>=0.0.21",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "authlib>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.1.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "python-ulid>=2.2.0",
    "email-validator>=2.1.0",
    "pydantic-settings>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
